+++ frontend/src/components/BadgeClassCard.vue
