+++ frontend/src/main.js
@@ -1,47 +1,13 @@
 import Vue from 'vue';
-import VueRouter from 'vue-router';
 import VueResource from 'vue-resource';
 
-import store from './store.js';
+import store from './store/index.js';
+import router from './router/index.js'
 
-import Landing from './Landing.vue';
-import Recipient from './Recipient.vue';
-import Share from './Share.vue';
 import App from './App.vue';
 
-Vue.use(VueRouter);
+
 Vue.use(VueResource);
-
-// Little helper for modeling the flow in the pages
-const flow = (pred, succ, redirect = '/') => (to, from, next) => {
-  if (from.name !== pred && from.name !== succ) {
-    return next(redirect);
-  }
-  return next();
-}
-
-const routes = [{
-    path: '/',
-    name: 'landing',
-    component: Landing,
-  },
-  {
-    path: '/recipient',
-    name: 'recipient',
-    component: Recipient,
-    props: true,
-    beforeEnter: flow('landing', 'share')
-  },
-  {
-    path: '/share',
-    name: 'share',
-    component: Share,
-    props: true,
-    beforeEnter: flow('recipient', undefined)
-  },
-];
-
-const router = new VueRouter({ routes });
 
 new Vue({
   router,
